# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZGalaxy - Planet & Moon Infrastructure Management API
  description: |
    ZGalaxy Planet & Moon Infrastructure Engine API. Manages self-hosted ZeroTier
    Planet/Moon infrastructure: authentication & RBAC, planet/moon building,
    federation mesh, cluster HA, Cloudflare DNS/DDNS, identity, and encrypted backup.
  version: 1.3.0
servers:
  - url: http://localhost:3000
    description: Local ZGalaxy server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
paths:
...
  /api/v1/cluster/nodes/add:
    post:
      summary: Add a cluster node
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nodeId, ip4]
              properties:
                nodeId: { type: string }
                name: { type: string }
                ip4: { type: string }
                ip6: { type: string }
                domain: { type: string }
                port: { type: integer }
                isLocal: { type: boolean }
      responses:
        '200':
          description: Node added
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/cluster/nodes/{nodeId}:
    delete:
      summary: Remove a cluster node
      security:
        - BearerAuth: []
      parameters:
        - name: nodeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Node removed
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/cluster/sync:
    post:
      summary: Re-probe node health (real checks) and sync topology
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Synced topology with honest ONLINE/OFFLINE status
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/cluster/build-unified:
    post:
      summary: Build a multi-root unified Planet from all ONLINE nodes
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Multi-root planet built
        '403':
          $ref: '#/components/schemas/Error'

  # ================= FEDERATION =================
  /api/v1/federation/handshake:
    post:
      summary: Inter-node handshake (public; token-gated, rate limited, anti-poisoning)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceNodeId, sourceEndpoint, tokenSecret]
              properties:
                sourceNodeId: { type: string }
                sourceNodeName: { type: string }
                sourceEndpoint: { type: string }
                tokenSecret: { type: string }
                requestedSyncMode:
                  type: string
                  enum: [FEDERATION_INHERITED, DIRECT_ISOLATED]
      responses:
        '200':
          description: Handshake accepted
        '400':
          $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
  /api/v1/federation/tokens:
    get:
      summary: List federation tokens (secrets masked)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token list with tokenSecretMasked only
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/federation/tokens/create:
    post:
      summary: Create a federation token (secret shown once)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                syncMode:
                  type: string
                  enum: [FEDERATION_INHERITED, DIRECT_ISOLATED]
                permissions:
                  type: array
                  items:
                    type: string
                    enum: [READ, WRITE, PLANET_SYNC, MOON_SYNC]
                maxUses: { type: integer }
                expiresInDays: { type: integer }
      responses:
        '201':
          description: Token created
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/federation/tokens/{id}/revoke:
    post:
      summary: Revoke a federation token
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Token revoked
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/federation/tokens/{id}/renew:
    post:
      summary: Extend a federation token expiration
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                extensionDays: { type: integer }
      responses:
        '200':
          description: Token renewed
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/federation/join:
    post:
      summary: Join a remote federation node via token (SSRF-guarded)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetEndpoint, tokenSecret]
              properties:
                targetEndpoint: { type: string }
                tokenSecret: { type: string }
                syncMode:
                  type: string
                  enum: [FEDERATION_INHERITED, DIRECT_ISOLATED]
      responses:
        '200':
          description: Joined
        '400':
          description: Invalid/blocked target (SSRF protection)
        '403':
          $ref: '#/components/schemas/Error'
...
  /api/v1/federation/peers/{nodeId}:
    delete:
      summary: Disconnect a peer
      security:
        - BearerAuth: []
      parameters:
        - name: nodeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Peer disconnected
        '404':
          $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/federation/sync-now:
    post:
      summary: Trigger real mesh topology propagation (fan-out)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Propagation results
        '403':
          $ref: '#/components/schemas/Error'

  # ================= IDENTITY =================
...
  /api/v1/domains/verify:
    post:
      summary: Verify a domain's DNS A/AAAA records
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain: { type: string }
      responses:
        '200':
          description: Verification result
  /api/v1/domains/bind:
    post:
      summary: Bind a domain to PLANET or MOON
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain: { type: string }
                target:
                  type: string
                  enum: [PLANET, MOON]
      responses:
        '200':
          description: Domain bound
        '403':
          $ref: '#/components/schemas/Error'

  # ================= DDNS =================
...